+++ packages/ui/src/index.ts
@@ -1,0 +1,7 @@
+export { Breadcrumbs } from './breadcrumbs';
+export { SummaryButton } from './summary-button';
+
+// WP Core (Unlocked private APIs)
+export { Badge } from './badge';
+export { Menu } from './menu';
+export { Tabs } from './tabs';